# cmd/docs/swagger.yaml
definitions:
...
  github_com_sankangkin_di-rest-api_internal_models.ItemTransaction:
    properties:
      createdAt:
        type: string
      createdTime:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        description: TODO to enhance with UUID
        type: integer
      inQty:
        type: integer
      outQty:
        type: integer
      productId:
        type: string
      referenceNo:
        type: string
      remark:
        type: string
      tranType:
        type: string
      uom:
        type: string
      updatedAt:
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  httputil.HttpError400:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  httputil.HttpError401:
    properties:
      code:
        example: 401
        type: integer
      message:
        example: Unauthorized
        type: string
    type: object
  httputil.HttpError500:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: internal server error
        type: string
    type: object
...
  internal_domain_itemtransactions.ResquestAdjustInventoryDTO:
    properties:
      baseQty:
        type: integer
      createdAt:
        description: Timestamp of the transaction
        type: string
      derivedQty:
        type: integer
      inQty:
        description: Quantity to be added
        type: integer
      outQty:
        description: Quantity to be removed
        type: integer
      productId:
        type: string
      referenceNo:
        description: Reference number for the transaction
        type: string
      remark:
        type: string
      tranType:
        description: DEBIT or CREDIT
        type: string
      uom:
        description: Unit of Measure (e.g., EACH, KG)
        type: string
    type: object
...
host: localhost:5555
info:
  contact:
    email: sankaungkin@gmail.com
    name: San KaungKin
  description: This is an auto-generated API docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: REST-API with(golang fiber, google wire dependency injection)
  version: "1.0"
paths:
...
  /api/transactions:
    get:
      consumes:
      - application/json
      description: Fetch all transactions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.ItemTransaction'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Fetch all transactions
      tags:
      - Transactions
  /api/transactions/adjustment:
    post:
      consumes:
      - application/json
      description: Create
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product input data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/internal_domain_itemtransactions.ResquestAdjustInventoryDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.ItemTransaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Create adjustment transaction
      tags:
      - Transactions
  /api/transactions/by-product-type/{productId}/{tranType}:
    get:
      consumes:
      - application/json
      description: Fetch individual transaction by productId and tranType
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      - description: Transaction Type
        in: path
        name: tranType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.ItemTransaction'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Fetch individual transaction by productId and tranType
      tags:
      - Transactions
  /api/transactions/by-product/{productId}:
    get:
      consumes:
      - application/json
      description: Fetch individual transaction by productId
      parameters:
      - description: product Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.ItemTransaction'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Fetch individual transaction by productId
      tags:
      - Transactions
  /api/transactions/by-type/{tranType}:
    get:
      consumes:
      - application/json
      description: Fetch individual transaction by protransactionType
      parameters:
      - description: transactionType
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.ItemTransaction'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Fetch individual transaction by transactionType
      tags:
      - Transactions
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"